back_end/requirements.txt
asyncpg==0.29.0
fastapi==0.111.0
pydantic-settings==2.2.1
python-dotenv==1.0.0
SQLAlchemy==1.4.47